Improve general UI, matches page, teams page

+++ src/routes/survey/teams/SurveyTeamsPage.svelte
@@ -3,9 +3,8 @@
   import Container from "$lib/components/Container.svelte";
   import Header from "$lib/components/Header.svelte";
   import Icon from "$lib/components/Icon.svelte";
-  import { tbaAuthKeyStore } from "$lib/settings";
+    import { modeStore } from "$lib/settings";
   import type { Survey } from "$lib/survey";
-  import { tbaGetEventTeams } from "$lib/tba";
 
   let {
     idb,
@@ -20,16 +19,6 @@
   });
 
   let teamInput = $state("");
-
-  async function getTeamsFromTBAEvent() {
-    if (!surveyRecord.tbaEventKey) return;
-
-    const response = await tbaGetEventTeams(surveyRecord.tbaEventKey, $tbaAuthKeyStore);
-    if (response) {
-      surveyRecord.teams = response;
-      surveyRecord.modified = new Date();
-    }
-  }
 
   function addTeam() {
     if (teamInput.trim() && !surveyRecord.teams.includes(teamInput.trim())) {
@@ -51,34 +40,31 @@
 </Header>
 
 <Container direction="column" padding="large">
-  {#if $tbaAuthKeyStore && surveyRecord.tbaEventKey}
-    <Button onclick={getTeamsFromTBAEvent}>
-      <Container maxWidth>
-        <Icon name="cloud-arrow-down" />
-        Get teams from TBA event: {surveyRecord.tbaEventKey}
+  {#if $modeStore == "admin"}
+    <Container direction="column" gap="none">
+      New team
+      <Container>
+        <input style="width:200px" bind:value={teamInput} onkeydown={(e) => e.key == "Enter" && addTeam()} />
+        <Button disabled={!teamInput.trim().length || surveyRecord.teams.includes(teamInput.trim())} onclick={addTeam}>
+          <Icon name="plus" />
+          Add
+        </Button>
       </Container>
-    </Button>
-  {/if}
-
-  <Container direction="column" gap="none">
-    Add team
-    <Container>
-      <input style="width:200px" bind:value={teamInput} onkeydown={(e) => e.key == "Enter" && addTeam()} />
-      <Button disabled={!teamInput.trim().length || surveyRecord.teams.includes(teamInput.trim())} onclick={addTeam}>
-        <Icon name="plus" />
-        Add
-      </Button>
     </Container>
-  </Container>
+  {/if}
 
   <h2>Teams</h2>
   {#if surveyRecord.teams.length}
     <Container>
-      {#each surveyRecord.teams.toSorted((a, b) => a.localeCompare(b, "en", { numeric: true })) as team}
-        <Button onclick={() => deleteTeam(team)}>
-          {team}
-        </Button>
-      {/each}
+      {#if $modeStore == "admin"}
+        {#each surveyRecord.teams.toSorted((a, b) => a.localeCompare(b, "en", { numeric: true })) as team}
+          <Button onclick={() => deleteTeam(team)}>
+            {team}
+          </Button>
+        {/each}
+      {:else}
+        {surveyRecord.teams.toSorted((a, b) => a.localeCompare(b, "en", { numeric: true })).join(", ")}
+      {/if}
     </Container>
   {:else}
     <span>
